Resolver avisos restantes do lint
