+++ src/components/public/RestaurantCard.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Card } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
 import { Star, Percent, Tag } from 'lucide-react';
