+++ src/components/reports/RestaurantRanking.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Star, TrendingUp, TrendingDown, DollarSign, ShoppingBag } from "lucide-react";
