+++ src/components/ui/badge.jsx
@@ -31,4 +31,4 @@ function Badge({
   return (<div className={cn(badgeVariants({ variant }), className)} {...props} />);
 }
 
-export { Badge, badgeVariants }
+export { Badge }
