+++ src/components/ui/sidebar.jsx
@@ -615,5 +615,4 @@ export {
   SidebarRail,
   SidebarSeparator,
   SidebarTrigger,
-  useSidebar,
 }
